Refactor the filenames and correlated paths
