+++ tests/antd4.test.tsx
@@ -3,6 +3,6 @@ import Button from "antd4/lib/button";
 import FormItem from "antd4/lib/form/FormItem";
 
 import commonTests from "./common";
-import PhoneInput from "../src/index4";
+import PhoneInput from "../src/Input4";
 
 commonTests(PhoneInput, Form, FormItem, Button);
