+++ tests/antd5.test.tsx
@@ -3,6 +3,6 @@ import Button from "antd/lib/button";
 import FormItem from "antd/lib/form/FormItem";
 
 import commonTests from "./common";
-import PhoneInput from "../src/index5";
+import PhoneInput from "../src/Input5";
 
 commonTests(PhoneInput, Form, FormItem, Button);
